cmake: Use GNUInstallDirs module to control install directories
Use the GNUInstallDirs module as a standard way to set install
directories.  It is available since CMake 2.8.5 (Jul 2011), and it
works the same across many CMake-based packages, providing a single
way to set and override install directories.  It also does libdir
detection automatically in a more reliable way.

That said, the previous detection code wrongly set 'LIBSUFFIX'
instead of 'LIB_SUFFIX', so it did not work at all.